Remove AbstractString inheritance
This was causing a problem with CSV reading. CSV reader uses
SentinelArrays, and that checks if the custom type you are asking to
parse inherits from AbstractString before passing to a custom parsing
mechanism optimized for strings. Unfortunately, this creates a String7
sentinel with eight \xff values, and String7 interprets this as a
length-255 string, which obviously cannot fit into a String7.

To fix this, the AbstractString interface was made more explicit. Rather
than just inheriting, the methods are manually defined, but only methods
that make sense. This drops many methods from implementation, so the API
change leads to a major version bump.